Improve QueJob#by_job_class performance.
This method wasn't making use of the que_jobs.args index and was
performing badly on large datasets.

GIN indexes don't support the '=' operator, but support existence '?'
and containment '@>' among others. To make use of the index the query
has to check for containment of a subset document rather than match a
value within the document directly.

QueJob#by_job_args is already querying in this way; this change makes
QueJob#by_job_class consistent with it.

Context:

https://www.postgresql.org/docs/9.6/datatype-json.html#JSON-INDEXING

Given a jsonb column "data" indexed as follows:
  CREATE INDEX data_index ON table USING GIN (data);

And with data in this format:
  {
    "name": "Melbourne",
    "type": "City"
  }

This query will make use of the index:
  SELECT data FROM table WHERE data @> '{"name": "Melbourne"}';

But this one will not:
  SELECT data FROM table WHERE data->>'name' = "Melbourne";